some changes

+++ database/migrations/2020_11_26_174816_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('Designation');
             $table->string('Type');
-            $table->integer('Quantity');
+            $table->double('Quantity');
             $table->double('Price_A');
             $table->double('Price_V');
