+++ database/migrations/2020_11_27_094135_create_commandes_table.php
@@ -19,7 +19,7 @@ class CreateCommandesTable extends Migration
             $table->string('phone');
             $table->string('Designation');
             $table->string('Type');
-            $table->integer('Quantity');
+            $table->double('Quantity');
            
             $table->timestamps();
         });
