+++ database/migrations/2020_11_28_191304_create_achats_table.php
@@ -16,7 +16,7 @@ class CreateAchatsTable extends Migration
         Schema::create('achats', function (Blueprint $table) {
             $table->id();
             $table->foreignId('product_id');
-            $table->integer('Quantity');
+            $table->double('Quantity');
             $table->double('Amount');
             $table->timestamps();
         });
